move amp middleware under amp directory
