+++ packages/frontend/app/server.tsx
@@ -1,6 +1,5 @@
 import * as path from 'path';
 import express from 'express';
-import React from 'react';
 
 import recordBaselineCloudWatchMetrics from './aws/metrics-baseline';
 import {
@@ -8,8 +7,6 @@ import {
     GuardianConfiguration,
 } from './aws/aws-parameters';
 import document from '@frontend/web/document';
-import AMPDocument from '@frontend/amp/document';
-import AMPArticle from '@frontend/amp/pages/Article';
 import { dist, root, port } from '@root/scripts/frontend/config';
 import { log, warn } from '@root/scripts/env/log';
 
@@ -17,8 +14,7 @@ import { extract as extractCAPI } from '@frontend/lib/model/extract-capi';
 import { extract as extractNAV } from '@frontend/lib/model/extract-nav';
 import { extract as extractGA } from '@frontend/lib/model/extract-ga';
 import { extract as extractConfig } from '@frontend/lib/model/extract-config';
-
-import { extractScripts } from '@frontend/amp/components/lib/AMPScripts';
+import { render as renderAMPArticle } from '@frontend/amp/server/render';
 
 const renderArticle = ({ body }: express.Request, res: express.Response) => {
     try {
@@ -31,26 +27,6 @@ const renderArticle = ({ body }: express.Request, res: express.Response) => {
                 config: extractConfig(body),
                 GA: extractGA(body),
             },
-        });
-
-        res.status(200).send(resp);
-    } catch (e) {
-        res.status(500).send(`<pre>${e.stack}</pre>`);
-    }
-};
-
-const renderAMPArticle = ({ body }: express.Request, res: express.Response) => {
-    try {
-        const CAPI = extractCAPI(body);
-        const resp = AMPDocument({
-            scripts: extractScripts(CAPI.elements),
-            body: (
-                <AMPArticle
-                    articleData={CAPI}
-                    nav={extractNAV(body)}
-                    config={extractConfig(body)}
-                />
-            ),
         });
 
         res.status(200).send(resp);
